requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
oauthlib==3.1.0
PySocks==1.7.1
python-dotenv==0.12.0
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
tweepy==3.8.0
urllib3==1.25.8
